hotfix: wmic deprecated (#402)
WMIC is deprecated as of Windows 10, version 21H1; and as of the 21H1
semi-annual channel release of Windows Server. This utility is
superseded by Windows PowerShell for WMI; see Chapter 7 - Working with
WMI. This deprecation applies only to the WMIC utility. Windows
Management Instrumentation (WMI) itself is not affected. Also see
Windows 10 features we're no longer developing.

https://learn.microsoft.com/en-us/windows/win32/wmisdk/wmic