fix: .tool-versions not working Nushell
The issue here was that the global environment was being inserted at the beginning of the MultiToolVersions slice, which mistakenly gave it a higher precedence than everything else in the slice. It should have a lowever precedence than the working directory, so the code has been modified to place it after that in the slice.

Fixes #420